Crate parquet_index

Binary that prints the [page index] of a parquet file

Install

parquet-layout can be installed using cargo:

cargo install parquet --features=cli

After this parquet-index should be available:

parquet-index XYZ.parquet COLUMN_NAME

The binary can also be built from the source code and run as follows:

cargo run --features=cli --bin parquet-index XYZ.parquet COLUMN_NAME

[page index]: https://github.com/apache/parquet-format/blob/master/PageIndex.md

Structs

Args 🔒

Functions

compute_row_counts 🔒

Computes the number of rows in each page within a column chunk

main 🔒

print_index 🔒

Prints index information for a single column chunk